RT-DETR_train_extraction_and_analysis/requirements.txt
coco-eval==0.0.4
hydra-core==1.3.2
mlflow==2.16.2
omegaconf==2.3.0
pycocotools==2.0.8
pytorch-lightning==2.4.0
torch==2.3.0
torchaudio==2.3.0
torchmetrics==1.4.2
torchvision==0.18.0
transformers==4.45.1
timm==1.0.14